Get the geometries of a Marine Regions Geo-Object

Get the geometries of a Marine Regions Geo-Object

Usage

gaz_geometry(x, ...)

# S3 method for class 'numeric'
gaz_geometry(x, ...)

# S3 method for class 'mr_df'
gaz_geometry(x, ...)

Arguments

x

object to retrieve the geometries from. Accepted:

(integer) A valid Marine Regions Gazetteer Identifier (MRGID)

A data frame retrieved with mregions2 via its functions gaz_search(), gaz_search_by_source(), gaz_search_by_type() or gaz_relations(). See details.

...

Arguments passed on to gaz_rest_geometries

format

(character) The preferred output format. One of:

"sfc": Simple Feature geometry object. See 'sf'

"wkt": Geometry representation as Well-Known Text

"rdf": Geometry as an object of class 'rdf". See 'rdflib'

Default is "sfc"

multipart

(logical) Some Geo-Objects are compound of more than one part.

If FALSE, returns singlepart geometries (e.g. POLYGON, LINESTRING)

If TRUE (default), returns multipart geometries (e.g. MULTIPOLYGON, MULTILINESTRING)

Value

A sfc object (default), a sf data frame, a WKT string or an RDF object

Details

You can pass the output of most gaz_* functions to gaz_geometry() to retrieve the geometry the gazetteer entry. The data frame is then transformed into a sf::sf object.

Developer info

This is done in the method gaz_geometry.mr_df(). mr_df is a class defined in this package to ensure the data frame passed to gaz_geometry has a variable with MRGID.

Examples

if (FALSE) { # \dontrun{
require(magrittr)

gaz_geometry(3293)
gaz_geometry(3293, format = "wkt")
gaz_geometry(3293, format = "rdf")

gaz_search(3293) %>% gaz_geometry()
} # }