Marine Regions Gazetteer RESTful services (Documentation)

RESTful service REST (REpresentational State Transfer) is a simple stateless architecture that generally runs over HTTP.

mregions2 makes use of the RESTful API created and maintained by Marine Regions. The functions with names starting as gaz_rest_* perform HTTP requests to read the Marine Regions REST API. They are closer to the definition of each function in the Marine Regions REST API. All the gazetteer functions such as gaz_search()or gaz_relations() make use of these gaz_rest_* functions.

Value

Returns a help page: gaz_rest is not a function but documentation.

See also

Examples

if (FALSE) ?gaz_rest