Get all the place types of the Marine Regions Gazetteer

Get all the place types of the Marine Regions Gazetteer

Usage

gaz_types()

Value

a data frame with three columns:

typeID: the identifier of the place type in the Marine Regions Gazetteer database.

type: the name of the place type.

description: if available, the description of the place type.

See also

Examples

# \donttest{
# This
gaz_rest_types()
#> # A tibble: 332 × 3
#>    typeID type                      description                                 
#>     <int> <chr>                     <chr>                                       
#>  1      1 Town                      The lowest administrative unit in Belgium a…
#>  2      2 Arrondissement            The administrative unit in Belgium that lie…
#>  3      3 Department                A high-level administrative unit that is us…
#>  4      4 Province (administrative) A medium-level administrative unit that is …
#>  5      5 Country                   A high-level administrative unit that only …
#>  6      6 Continent                 The highest subdivision of the world. There…
#>  7      7 Region                    A high-level administrative unit that is us…
#>  8      8 Ward                      The lowest administrative unit in the Unite…
#>  9      9 Commune                   The lowest administrative unit in a lot of …
#> 10     10 District                  A medium-level administrative unit that is …
#> # ℹ 322 more rows

# is the same as
gaz_types()
#> # A tibble: 332 × 3
#>    typeID type                      description                                 
#>     <int> <chr>                     <chr>                                       
#>  1      1 Town                      The lowest administrative unit in Belgium a…
#>  2      2 Arrondissement            The administrative unit in Belgium that lie…
#>  3      3 Department                A high-level administrative unit that is us…
#>  4      4 Province (administrative) A medium-level administrative unit that is …
#>  5      5 Country                   A high-level administrative unit that only …
#>  6      6 Continent                 The highest subdivision of the world. There…
#>  7      7 Region                    A high-level administrative unit that is us…
#>  8      8 Ward                      The lowest administrative unit in the Unite…
#>  9      9 Commune                   The lowest administrative unit in a lot of …
#> 10     10 District                  A medium-level administrative unit that is …
#> # ℹ 322 more rows
# }