Retrieve Gazetteer Records by Placetype

Every record in the [Marine Gazetteer](https://marineregions.org/gazetteer.php) has a placetype, e.g. `Sandbank` or `Marine Protected Area`. All placetypes currently available in the gazetteer can be retrieved via `mr_gaz_info("placetypes")`.

Usage

mr_gaz_records_by_type(type, offset = 0)

Arguments

type

The placetype of which records are retrieved. Must be one of the types in `mr_gaz_info("placetypes")` and class `character`. Case insensitive.

offset

Start record number, in order to page through next batch of results.

Value

A tibble with all Gazetteer records of the specified placetype.

Examples

fao_subdivs <- mr_gaz_records_by_type("FAO Subdivisions")
fao_subdivs$preferredGazetteerName[1]
#> [1] "Northern Alboran Sea"
# [1] "Northern Alboran Sea"