Retrieve Gazetter Relations by MRGID

Gazetteer Records that are related to a given input MRGID. Relationships can be parents (`upper`), children (`lower`) oder both. Geographic types can also be specified, for example `partof` and `adjacentto`.

Usage

mr_gaz_relations_by_MRGID(mrgid, direction = "upper", type = "partof")

Arguments

mrgid

The [Marine Regions Geographic IDentifier](https://marineregions.org/mrgid.php).

direction

The hierarchical structure. Must be one of `c("upper", "lower", "both")`. `"upper"` lists all parents of the record. `"lower"` lists all childs of the record. `"both"` lists parents and childs of the record.

type

must be one of `c("partof", "partlypartof", "adjacentto", "similarto", "administrativepartof", "influencedby", "all")`. Explanations of the `types` [here](https://marineregions.org/ontology/documentation.html) in chapter *Object Properties*.

Value

a `tibble` with all relations for the given `MRGID`.

Examples

mariana_trench <- mr_gaz_records_by_names("Mariana Trench")
mariana_trench_mrgid <- mariana_trench$MRGID

mariana_trench_relations <- mr_gaz_relations_by_MRGID(mariana_trench_mrgid)
mariana_trench_relations$preferredGazetteerName
#> [1] "North Pacific Ocean"