Get a data product

Get a data product

Usage

mrp_get(product_name, ...)

Arguments

product_name

(character) Identifier of the data product. See mrp_list()

...

Any vendor params to be passed to the GetFeature request. See details.

Value

An sf object with the Marine Regions data product

Details

This function uses WFS services to load retrieve the Marine Regions Data products. The data products are stored in a geoserver instance hosted and maintained by the Flanders Marine Institute (VLIZ)

Vendor Params

Vendor params can be passed to a WFS GetFeature request. mregions2 uses the ows4R package as the interface to query the WFS service with R. You can pass any vendor parameters via .... For instance, you can pass CQL or OGC filters with cql_filter="<the filter>" or filter="<the filter>". You can change the projection with srsName="<CRS>", only the features in a certain bounding box with bbox="a1,b1,a2,b2,<CRS>" where a1, b1, a2 and b2 are the coordinates and the CRS selects the coordinate reference system of this bounding box, or any other vendor parameters accepted by geoserver.

Filtering

Both the Contextual Query Language (CQL) filter and the standard OGC filter specification allow to query the server before performing a request. This will boost performance as you will only retrieve the area of your interest. It is possible to query on attributes, but also perform geospatial queries. For instance, you can query a bounding box of interest.

A tutorial on CQL filters is available in the geoserver web site.

See also

Examples

if (FALSE) {
# See the list of all data products
View(mrp_list())

# We want the Exclusive Economic Zones of Portugal. Let's first visualize the product:
mrp_view("eez")

# See all the columns on this data product
mrp_colnames("eez")

# We should query on sovereign
# See all the possible values of sovereign1, sovereign2 and sovereign3
sov1 = mrp_col_unique("eez", "sovereign1")
sov2 = mrp_col_unique("eez", "sovereign2")
sov3 = mrp_col_unique("eez", "sovereign3")

# Is Portugal a value in the sovereign1, 2 and 3?
"Portugal" %in% sov1
#> [1] TRUE

"Portugal" %in% sov2
#> [1] FALSE

"Portugal" %in% sov3
#> [1] FALSE

# Portugal is only in sovereign1. Let's write a CQL filter to get only
# the EEZs of Portugal, or those where Portugal is a party of a dispute or a joint regime
portugal_eez <- mrp_get("eez", cql_filter = "sovereign1 = 'Portugal'")

# You can also limit the number of features to be requested
mrp_get("eez", count = 5)
}