Visualize a Marine Regions data product without downloading.

Visualize a Marine Regions data product without downloading.

A series of helpers are available to ease the selection of the data products. Example: instead of running

mrp_view("eez")

You can use

mrp_view_eez()

Try mrp_view_*()with the identifier of the data product (see mrp_list)

Usage

mrp_view(layer, cql_filter = NULL, filter = NULL)

mrp_view_eez(...)

mrp_view_eez_boundaries(...)

mrp_view_eez_12nm(...)

mrp_view_eez_24nm(...)

mrp_view_eez_internal_waters(...)

mrp_view_eez_archipelagic_waters(...)

mrp_view_high_seas(...)

mrp_view_ecs(...)

mrp_view_ecs_boundaries(...)

mrp_view_iho(...)

mrp_view_goas(...)

mrp_view_eez_iho(...)

mrp_view_eez_land(...)

mrp_view_longhurst(...)

mrp_view_cds(...)

mrp_view_eca_reg13_nox(...)

mrp_view_eca_reg14_sox_pm(...)

mrp_view_worldheritagemarineprogramme(...)

mrp_view_lme(...)

mrp_view_ecoregions(...)

mrp_view_seavox_v18(...)

Arguments

layer

(character) Identifier of the data product. See mrp_list

cql_filter

(character) Contextual Query Language (CQL) filter. See details.

filter

(character) Standard OGC filter specification. See details.

...

pass the cql_filter and filter parameters to mrp_view() when using one of the helpers

Value

A leaflet map with a data product visualized via WMS

Details

This function uses WMS services to load quickly a Leaflet viewer of a Marine Regions data product. It uses the EMODnet Bathymetry Digital Terrain Model as background layer.

Filters

Both the Contextual Query Language (CQL) filter and the standard OGC filter specification allow to query the server before performing a request. This will boost performance as you will only retrieve the area of your interest. It is possible to query on attributes, but also perform geospatial queries. For instance, you can query a bounding box of interest.

CQL filters are possible only in geoserver. Marine Regions uses a geoserver instance to serve its data products. A tutorial on CQL filters is available in the geoserver web site.

See also

Examples

# \donttest{
# You can pass a product name from mrp_list
mrp_view('eez')

# Or use the helper
mrp_view_eez()

# Example: filter a the Ecoregions 'Azores Canaries Madeira' with mrgid 21885
# You can check the names of the columns beforehand with mrp_colnames('ecoregions')
mrp_view_ecoregions(filter = "
  <Filter>
    <PropertyIsEqualTo>
      <PropertyName>ecoregion</PropertyName>
      <Literal>Azores Canaries Madeira</Literal>
    </PropertyIsEqualTo>
  </Filter>
")

# OGC filter are very verbose... but luckily you can use a CQL filter instead
mrp_view_ecoregions(cql_filter = "ecoregion = 'Azores Canaries Madeira'")

# View all the Extended Continental Shelf (ECS) boundary lines published during the first
# decade of the 21st century
mrp_view_ecs_boundaries(
  cql_filter = "doc_date > '2000-01-01' AND doc_date < '2009-12-31'"
)

# Or as timestamp
mrp_view_eez_boundaries(
  cql_filter = "doc_date AFTER 2000-01-01T00:00:00Z AND doc_date BEFORE 2009-12-31T00:00:00Z"
)

# }